slugify introduced

+++ post/views.py
@@ -1,9 +1,8 @@
 from django.contrib import messages
-from django.shortcuts import render, get_object_or_404, HttpResponseRedirect, redirect
+from django.shortcuts import render, get_object_or_404, HttpResponseRedirect, redirect, Http404
 
 from .forms import PostForm
 from .models import Post
-
 
 # Create your views here.
 
@@ -13,8 +12,9 @@ def post_index(request):
     return render(request, 'post/index.html', {'postlar': posts})
 
 
-def post_detail(request, id):
-    post = get_object_or_404(Post, id=id)
+def post_detail(request, slug):
+
+    post = get_object_or_404(Post, slug=slug)
     context = {
         'post': post,
     }
@@ -22,6 +22,9 @@ def post_detail(request, id):
 
 
 def post_create(request):
+    if not request.user.is_authenticated:
+        raise Http404
+
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         post = form.save()
@@ -35,12 +38,15 @@ def post_create(request):
     return render(request, 'post/form.html', context)
 
 
-def post_update(request, id):
-    post = get_object_or_404(Post, id=id)
+def post_update(request, slug):
+    if not request.user.is_authenticated:
+        raise Http404
+
+    post = get_object_or_404(Post, slug=slug)
     # GET ile gelindiğinde sadece instance=post bilgileri ile form nesnesi oluşturulur
     # POST ile gelindiğinde istek ile gelen form bilgilerindeki parametreler ile form nesnesi oluşturulur, instance
     # verildği için de form bunu changed data ile tutarak güncelleme yapacağını anlar. Instance verilmez ise create olur
-    form = PostForm(request.POST or None,  request.FILES or None, instance=post)
+    form = PostForm(request.POST or None, request.FILES or None, instance=post)
     if form.is_valid():
         post = form.save()
         messages.success(request, 'Kayıt başarılı bir şekilde güncellendi')
@@ -53,8 +59,11 @@ def post_update(request, id):
     return render(request, 'post/form.html', context)
 
 
-def post_delete(request, id):
-    post = get_object_or_404(Post, id=id)
+def post_delete(request, slug):
+    if not request.user.is_authenticated:
+        raise Http404
+
+    post = get_object_or_404(Post, slug=slug)
     form = PostForm(instance=post)
     if request.method == 'POST':
         post.delete()
